use configurable logger
