--- api/cluster_test.go
@@ -1,9 +1,0 @@
-package api
-
-import (
-	"testing"
-)
-
-func TestCreateCluster(t *testing.T)  {
-
-}
